Clean up after extracting backends
