+++ lib/delayed/worker.rb
@@ -2,6 +2,7 @@ require 'timeout'
 require 'active_support/core_ext/numeric/time'
 require 'active_support/core_ext/class/attribute_accessors'
 require 'active_support/core_ext/kernel'
+require 'logger'
 
 module Delayed
   class Worker
