+++ spec/sample_jobs.rb
@@ -29,3 +29,33 @@ module M
     def perform; @@runs += 1; end    
   end
 end
+
+class SuccessfulCallbackJob
+  cattr_accessor :messages
+
+  def before(job)
+    SuccessfulCallbackJob.messages << 'before perform'
+  end
+  
+  def perform
+    SuccessfulCallbackJob.messages << 'perform'
+  end
+  
+  def after(job, error = nil)
+    SuccessfulCallbackJob.messages << 'after perform'
+  end
+  
+  def success(job)
+    SuccessfulCallbackJob.messages << 'success!'
+  end
+  
+  def failure(job, error)
+    SuccessfulCallbackJob.messages << "error: #{error.class}"
+  end
+end
+
+class FailureCallbackJob < SuccessfulCallbackJob
+  def perform
+     raise "failure job"
+  end
+end
