--- spec/setup/active_record.rb
@@ -1,33 +1,0 @@
-require 'active_record'
-
-ActiveRecord::Base.establish_connection(:adapter => 'sqlite3', :database => ':memory:')
-ActiveRecord::Base.logger = Delayed::Worker.logger
-ActiveRecord::Migration.verbose = false
-
-ActiveRecord::Schema.define do
-  create_table :delayed_jobs, :force => true do |table|
-    table.integer  :priority, :default => 0
-    table.integer  :attempts, :default => 0
-    table.text     :handler
-    table.text     :last_error
-    table.datetime :run_at
-    table.datetime :locked_at
-    table.datetime :failed_at
-    table.string   :locked_by
-    table.timestamps
-  end
-
-  add_index :delayed_jobs, [:priority, :run_at], :name => 'delayed_jobs_priority'
-
-  create_table :stories, :force => true do |table|
-    table.string :text
-  end
-end
-
-# Purely useful for test cases...
-class Story < ActiveRecord::Base
-  def tell; text; end       
-  def whatever(n, _); tell*n; end
-  
-  handle_asynchronously :whatever
-end
