+++ spec/spec_helper.rb
@@ -7,24 +7,46 @@ require 'logger'
 gem 'activerecord', ENV['RAILS_VERSION'] if ENV['RAILS_VERSION']
 
 require 'delayed_job'
+require 'delayed/backend/shared_spec'
 
 Delayed::Worker.logger = Logger.new('/tmp/dj.log')
 RAILS_ENV = 'test'
 
-# determine the available backends
-BACKENDS = []
-Dir.glob("#{File.dirname(__FILE__)}/setup/*.rb") do |backend|
-  begin
-    backend = File.basename(backend, '.rb')
-    require "setup/#{backend}"
-    require "backend/#{backend}_job_spec"
-    BACKENDS << backend.to_sym
-  rescue LoadError
-   	puts "Unable to load #{backend} backend: #{$!}"
+require 'active_record'
+
+ActiveRecord::Base.establish_connection(:adapter => 'sqlite3', :database => ':memory:')
+ActiveRecord::Base.logger = Delayed::Worker.logger
+ActiveRecord::Migration.verbose = false
+
+ActiveRecord::Schema.define do
+  create_table :delayed_jobs, :force => true do |table|
+    table.integer  :priority, :default => 0
+    table.integer  :attempts, :default => 0
+    table.text     :handler
+    table.text     :last_error
+    table.datetime :run_at
+    table.datetime :locked_at
+    table.datetime :failed_at
+    table.string   :locked_by
+    table.timestamps
+  end
+
+  add_index :delayed_jobs, [:priority, :run_at], :name => 'delayed_jobs_priority'
+
+  create_table :stories, :force => true do |table|
+    table.string :text
   end
 end
 
-Delayed::Worker.backend = BACKENDS.first
+# Purely useful for test cases...
+class Story < ActiveRecord::Base
+  def tell; text; end       
+  def whatever(n, _); tell*n; end
+  
+  handle_asynchronously :whatever
+end
+
+Delayed::Worker.backend = :active_record
 
 # Add this directory so the ActiveSupport autoloading works
 ActiveSupport::Dependencies.autoload_paths << File.dirname(__FILE__)
